+++ Kernel/Modules/AdminCustomerCompany.pm
@@ -50,6 +50,8 @@ sub Run {
         my %Data = $Self->{CustomerCompanyObject}->CustomerCompanyGet(
             CustomerID => $CustomerID,
         );
+        $Data{CustomerCompanyID} = $CustomerID;
+
         my $Output = $Self->{LayoutObject}->Header();
         $Output .= $Self->{LayoutObject}->NavigationBar(
             Type => $NavigationBarType,
@@ -78,6 +80,8 @@ sub Run {
         my $Note = '';
         my ( %GetParam, %Errors );
 
+        $GetParam{Source} = $Self->{ParamObject}->GetParam( Param => 'Source' );
+
         $GetParam{CustomerCompanyID} = $Self->{ParamObject}->GetParam( Param => 'CustomerCompanyID' );
 
         for my $Entry ( @{ $Self->{ConfigObject}->Get('CustomerCompany')->{Map} } ) {
@@ -91,6 +95,21 @@ sub Run {
 
         if ( !defined $GetParam{CustomerID} ) {
             $GetParam{CustomerID} = $Self->{ParamObject}->GetParam( Param => 'CustomerID' ) || '';
+        }
+
+        # check for duplicate entries
+        if ( $GetParam{CustomerCompanyID} ne $GetParam{CustomerID} ) {
+
+            # get CustomerCompany list
+            my %List = $Self->{CustomerCompanyObject}->CustomerCompanyList(
+                Search => $Param{Search},
+                Valid  => 0,
+            );
+
+            # check duplicate field
+            if ( %List && $List{ $GetParam{CustomerID} } ) {
+                $Errors{Duplicate} = 'ServerError';
+            }
         }
 
         # if no errors occurred
@@ -127,6 +146,15 @@ sub Run {
             Type => $NavigationBarType,
         );
         $Output .= $Self->{LayoutObject}->Notify( Priority => 'Error' );
+
+        # set notification for duplicate entry
+        if ( $Errors{Duplicate} ) {
+            $Output .= $Self->{LayoutObject}->Notify(
+                Priority => 'Error',
+                Info     => "CustomerCompany $GetParam{CustomerID} already exists!.",
+            );
+        }
+
         $Self->_Edit(
             Action => 'Change',
             Nav    => $Nav,
@@ -175,6 +203,11 @@ sub Run {
 
         my $Note = '';
         my ( %GetParam, %Errors );
+
+        $GetParam{Source} = $Self->{ParamObject}->GetParam( Param => 'Source' );
+        my $CustomerCompanyKey = $Self->{ConfigObject}->Get('CustomerCompany')->{CustomerCompanyKey};
+        my $CustomerCompanyID;
+
         for my $Entry ( @{ $Self->{ConfigObject}->Get('CustomerCompany')->{Map} } ) {
             $GetParam{ $Entry->[0] } = $Self->{ParamObject}->GetParam( Param => $Entry->[0] ) || '';
 
@@ -182,6 +215,22 @@ sub Run {
             if ( !$GetParam{ $Entry->[0] } && $Entry->[4] ) {
                 $Errors{ $Entry->[0] . 'Invalid' } = 'ServerError';
             }
+
+            # save customer company key for checking duplicate
+            if ( $Entry->[2] eq $CustomerCompanyKey ) {
+                $CustomerCompanyID = $GetParam{ $Entry->[0] };
+            }
+        }
+
+        # get CustomerCompany list
+        my %List = $Self->{CustomerCompanyObject}->CustomerCompanyList(
+            Search => $Param{Search},
+            Valid  => 0,
+        );
+
+        # check duplicate field
+        if ( %List && $List{$CustomerCompanyID} ) {
+            $Errors{Duplicate} = 'ServerError';
         }
 
         # if no errors occurred
@@ -218,6 +267,15 @@ sub Run {
             Type => $NavigationBarType,
         );
         $Output .= $Self->{LayoutObject}->Notify( Priority => 'Error' );
+
+        # set notification for duplicate entry
+        if ( $Errors{Duplicate} ) {
+            $Output .= $Self->{LayoutObject}->Notify(
+                Priority => 'Error',
+                Info     => "CustomerCompany $GetParam{CustomerID} already exists!.",
+            );
+        }
+
         $Self->_Edit(
             Action => 'Add',
             Nav    => $Nav,